Using ptrdiff_t instead of int for difference_type.

+++ include/garlic/clove.h
@@ -95,7 +95,7 @@ namespace garlic {
       ValueType value;
     };
 
-    using difference_type = int;
+    using difference_type = ptrdiff_t;
     using value_type = MemberWrapper;
     using reference = ValueType&;
     using pointer = ValueType*;
