+++ include/garlic/providers/yaml-cpp.h
@@ -22,7 +22,7 @@ namespace garlic::providers::yamlcpp {
       ValueType value;
     };
     
-    using difference_type = int;
+    using difference_type = ptrdiff_t;
     using value_type = MemberWrapper;
     using reference = MemberWrapper&;
     using pointer = MemberWrapper*;
